Wit.ai Speech to Text for Question Answering (#146)
Add a toolbar icon - when clicked, it will open a small window that allows user to record audio. The audio is recorded in WAV format and sent to the wit.ai speech API, which converts speech to text. Once the text is returned we call the Flask /answer endpoint to get a result for the query.