Indexer refactoring: creation of StringTransform (#161)
Refinement of the object model for the indexing pipeline. A StringTransform is a string-to-string transformation. In the context of the indexing pipeline, it can be used to, for example, clean HTML documents into plain text documents.

Previously, this functionality was tucked away inside the LuceneDocumentGenerator - so that if we wanted it explore different HTML parsers, for example, we would have needed to duplicate a lot of code. Now, LuceneDocumentGenerator has pluggable StringTransforms.

This allowed us to easily run a long-standing experiment I've always wondered about: Jsoup vs. Lucene's use of the NekoHTML parser.