POC of simplified regression test script and usage (#286)
In this PR, a new script `src/main/python/run_regression.py` and the expected results in `src/main/resources/regression/all.yaml` are introduced. 
This script is going to be the single regression test script for all collections (via different parameters). 
The idea is to separate "what to test" (in yaml file) and "how to test" (the script).
Currently, only the results of disk12 and cw09 are included but the initial POC seems promising. 
I temporarily name all current `experiments-*.md` as `experiments-*-old.md` and change the corresponding regression test scripts in `src/main/python/run_regression_*.py`
We will migrate all other collections to the new regression test script in another PR.