Allow stopwords to be configured at search time (#1473)
Stopwords can currently be configured at index-time. This PR adds support for
specifying stopwords at search-time as well.

This is helpful because generally the same stopwords are used at both index and
search time. If the stopword 'where' is dropped from the index but still
included in searches, then a search for 'where' only matches words with the same
stem like 'wheres'. This is tricky behavior and usually undesirable.