Add intra-run multithreading to SearchCollection (#1678)
Context: SearchCollection currently has the ability to do parameter sweeps, with different 
configurations in parallel using a thread pool. However, it currently has no ability to run 
different queries in a single configuration in parallel. This means that some sparse models 
take a long time to run.

This patch adds a thread pool within an individual configuration to run queries in parallel.