Add run_regressions_with_load.py (#2029)
This is a driver script for running a bunch of regressions (taken from a text file). Loops over the regression
commands and will fire off new jobs while the system load is below a threshold. I wrote this as a utility to
run a bunch of regression tests without overloading the system.