adding to reproduction log
I performed the reproduction on linux (ubuntu 22.04.4 LTS) from terminal. During the reproduction process, I was able to follow all the steps in the guide except for the python scripts, which I ran with 'python3' instead of 'python' as the command. The checks provided in the guide showed that this led to the same results.