Adjust IndexInfo naming to align with CLI arguments (#2736)
Also make class names direct references instead of strings
Two usages are supported: -encoder CosDprDistil OR -encoder CosDprDistilEncoder work. Previously it was the former usage.
Lastly, add tests for them as well. 

However, SearchHnswDenseVectorsTest is missing a test for this behaviour, because it causes a downstream failure at ArcticEmbedLEncoder. Currently, we don't know why this happens, commits / CI in the original PR show various experiments with this.